+++ src/aks-preview/azext_aks_preview/_consts.py
@@ -30,7 +30,7 @@ CONST_DEFAULT_NODE_OS_TYPE = "Linux"
 # os sku
 CONST_OS_SKU_UBUNTU = "Ubuntu"
 CONST_OS_SKU_CBLMARINER = "CBLMariner"
-CONST_OS_SKU_CBLMARINERV2 = "CBLMarinerV2"
+CONST_OS_SKU_MARINER = "Mariner"
 CONST_OS_SKU_WINDOWS2019 = "Windows2019"
 CONST_OS_SKU_WINDOWS2022 = "Windows2022"
 
@@ -184,3 +184,8 @@ CONST_PERISCOPE_NAMESPACE = "aks-periscope"
 
 CONST_AZURE_KEYVAULT_NETWORK_ACCESS_PUBLIC = "Public"
 CONST_AZURE_KEYVAULT_NETWORK_ACCESS_PRIVATE = "Private"
+
+# refer https://api.github.com/repos/Azure/draft/releases/latest
+# tag_name gives latest version released.
+# Moving away from 1:n release to avoid unwanted breaking changes with auto upgrades.
+CONST_DRAFT_CLI_VERSION = "v0.0.22"