+++ src/containerapp-compose/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+0.2.2
+++++++
+* Redirect `--transport` to `--transport-mapping` in preparation to move the command to the `containerapp` extension
+
+0.2.1
+++++++
+* Fix bug where location parameter was not being passed to containerapp environment creation
+* Make memory/cpu resolution more resilient and improve warning messages
+
 0.2.0
 ++++++
 * Add basic build support for compose services