+++ src/network-manager/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.5.3
++++++
+* Fix force delete behavior.
+
 0.5.2
 +++++
 * 'az network manager connection management-group': reopen the set of commands.